[DAY6-171] refactor : 최저가 계산 로직 PlaceController로 이동
